server/app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorlog==6.10.1
datasets==4.7.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dspy==2.5.17
dspy-ai==2.5.17
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.78.0
grpcio-tools==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json_repair==0.58.5
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm==1.49.1
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
numpy==2.4.3
openai==1.51.0
optuna==4.7.0
packaging==26.0
pandas==3.0.1
portalocker==2.10.1
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
qdrant-client==1.9.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
structlog==25.5.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
